move SplatTokens class into module
